back-end/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==4.2.1
    # via -r requirements.in
blinker==1.8.2
    # via
    #   -r requirements.in
    #   flask
click==8.1.8
    # via
    #   -r requirements.in
    #   flask
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-jwt-extended
    #   flask-sqlalchemy
flask-cors==5.0.0
    # via -r requirements.in
flask-jwt-extended==4.6.0
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.1.1
    # via
    #   -r requirements.in
    #   sqlalchemy
importlib-metadata==8.5.0
    # via
    #   -r requirements.in
    #   flask
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.5
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   flask-jwt-extended
python-dotenv==1.0.1
    # via -r requirements.in
sqlalchemy==2.0.37
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   sqlalchemy
werkzeug==3.0.6
    # via
    #   -r requirements.in
    #   flask
    #   flask-jwt-extended
zipp==3.20.2
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
